+++ examples/data-sources/nxos_nve_vni/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_nve_vni" "example" {
-  vni = 103100
-}