+++ examples/data-sources/nxos_nve_vni_container/data-source.tf
@@ -1,2 +1,0 @@
-data "nxos_nve_vni_container" "example" {
-}