+++ examples/data-sources/nxos_nve_vni_ingress_replication/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_nve_vni_ingress_replication" "example" {
-  vni = 103100
-}