+++ examples/resources/nxos_nve_vni/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_nve_vni.example
-  identity = {
-    "vni" : "<vni>"
-  }
-}