+++ examples/resources/nxos_nve_vni/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_nve_vni.example
-  id = "<vni>"
-}