+++ examples/resources/nxos_nve_vni_container/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_nve_vni_container.example
-  id = ""
-}