+++ examples/resources/nxos_nve_vni_container/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_nve_vni_container.example ""