+++ examples/resources/nxos_nve_vni_container/resource.tf
@@ -1,2 +1,0 @@
-resource "nxos_nve_vni_container" "example" {
-}