+++ examples/resources/nxos_nve_vni_ingress_replication/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_nve_vni_ingress_replication.example
-  id = "<vni>"
-}