+++ examples/resources/nxos_nve_vni_ingress_replication/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_nve_vni_ingress_replication.example "<vni>"