+++ examples/resources/nxos_nve_vni_ingress_replication/resource.tf
@@ -1,4 +1,0 @@
-resource "nxos_nve_vni_ingress_replication" "example" {
-  vni      = 103100
-  protocol = "bgp"
-}